+++ src/hooks/index.js
@@ -1,1 +1,2 @@
 export { default as useScript } from './useScript'
+export { default as useNormalizePhone} from "./useNormalizePhone"
